+++ src/signed-source/src/__tests__/SignedSource.test.js
@@ -5,11 +5,11 @@ import os from 'os';
 import SignedSource from '../SignedSource';
 
 test('signFile', () => {
-  expect(SignedSource.signFile(`# ${SignedSource.getSigningToken()}\ntest 1`)).toEqual(
+  expect(SignedSource.signFile(`# ${SignedSource.getSigningToken()}\ntest 1`)).toBe(
     `# @generated SignedSource<<d9b7b52f54978f54b84a0fd48145e470>>${os.EOL}test 1`,
   );
 
-  expect(SignedSource.signFile(`# ${SignedSource.getSigningToken()}\ntest 2`)).toEqual(
+  expect(SignedSource.signFile(`# ${SignedSource.getSigningToken()}\ntest 2`)).toBe(
     `# @generated SignedSource<<4c0c1ae4f5863c72731b2f543e830fd5>>${os.EOL}test 2`,
   );
 
@@ -18,7 +18,7 @@ test('signFile', () => {
     SignedSource.signFile(
       `# @generated SignedSource<<eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee>>\nalready signed test`,
     ),
-  ).toEqual(`# @generated SignedSource<<54e8ffafff15a19f858d95c9a13d5b1d>>\nalready signed test`);
+  ).toBe(`# @generated SignedSource<<54e8ffafff15a19f858d95c9a13d5b1d>>\nalready signed test`);
 
   expect(() =>
     SignedSource.signFile(`signature missing, no sign token`),